refactoring challenge, ranking
