+++ drivinganalysis/src/main/java/com/sejong/drivinganalysis/dto/challengedto/UserChallengeJoinRequest.java
@@ -1,4 +1,4 @@
-package com.sejong.drivinganalysis.challenge.dto;
+package com.sejong.drivinganalysis.dto.challengedto;
 
 import jakarta.validation.constraints.NotNull;
 import lombok.Getter;
