+++ drivinganalysis/src/main/java/com/sejong/drivinganalysis/service/UserChallengeService.java
@@ -1,8 +1,8 @@
-package com.sejong.drivinganalysis.challenge;
+package com.sejong.drivinganalysis.service;
 
-import com.sejong.drivinganalysis.challenge.dto.UserChallengeCreateRequest;
-import com.sejong.drivinganalysis.challenge.dto.UserChallengeResponse;
-import com.sejong.drivinganalysis.challenge.dto.UserChallengeJoinRequest;
+import com.sejong.drivinganalysis.dto.challengedto.UserChallengeCreateRequest;
+import com.sejong.drivinganalysis.dto.challengedto.UserChallengeResponse;
+import com.sejong.drivinganalysis.dto.challengedto.UserChallengeJoinRequest;
 import com.sejong.drivinganalysis.entity.User;
 import com.sejong.drivinganalysis.entity.UserChallenge;
