+++ drivinganalysis/src/main/resources/application.yml
@@ -2,8 +2,8 @@
 spring:
   datasource:
     url: jdbc:mysql://driving-analysis2.c5us24cq2762.ap-northeast-2.rds.amazonaws.com:3306/driving_analysis?useSSL=false&serverTimezone=UTC&characterEncoding=UTF-8&allowPublicKeyRetrieval=true
-    username:
-    password:
+    username: admin
+    password: whdals226!
     driver-class-name: com.mysql.cj.jdbc.Driver
 
   jpa:
@@ -36,7 +36,7 @@ logging:
 ai:
   server:
     host: 0.tcp.jp.ngrok.io
-    port: 12364
+    port: 13384
     timeout: 30  # 요청 타임아웃(초)
     retry:
       max: 2     # 최대 재시도 횟수
